extract TabbedComponentHelper to share tab rules between TabNav and UnderlineNav
