+++ test/components/tab_nav_component_test.rb
@@ -5,8 +5,16 @@ require "test_helper"
 class PrimerTabNavComponentTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
+  def test_raises_if_no_tab_is_rendered
+    err = assert_raises Primer::TabbedComponentHelper::NoSelectedTabsError do
+      render_inline(Primer::TabNavComponent.new)
+    end
+
+    assert_equal("a tab must be selected", err.message)
+  end
+
   def test_raises_if_no_tab_is_selected
-    err = assert_raises Primer::TabNavComponent::NoSelectedTabsError do
+    err = assert_raises Primer::TabbedComponentHelper::NoSelectedTabsError do
       render_inline(Primer::TabNavComponent.new) do |c|
         c.tab { "Tab 1" }
         c.tab { "Tab 2" }
@@ -18,7 +26,7 @@ class PrimerTabNavComponentTest < Minitest::Test
   end
 
   def test_raises_if_multiple_tabs_are_selected
-    err = assert_raises Primer::TabNavComponent::MultipleSelectedTabsError do
+    err = assert_raises Primer::TabbedComponentHelper::MultipleSelectedTabsError do
       render_inline(Primer::TabNavComponent.new) do |c|
         c.tab(selected: true) do
           "Tab 1"
