+++ test/components/underline_nav_component_test.rb
@@ -5,14 +5,43 @@ require "test_helper"
 class PrimerUnderlineNavComponentTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
-  def test_does_not_render_without_tabs
-    render_inline(Primer::UnderlineNavComponent.new) do |component|
-      component.actions do
-        "Actions content"
+  def test_raises_if_no_tab_is_rendered
+    err = assert_raises Primer::TabbedComponentHelper::NoSelectedTabsError do
+      render_inline(Primer::UnderlineNavComponent.new) do |component|
+        component.actions do
+          "Actions content"
+        end
       end
     end
 
-    refute_component_rendered
+    assert_equal("a tab must be selected", err.message)
+  end
+  def test_raises_if_no_tab_is_selected
+    err = assert_raises Primer::TabbedComponentHelper::NoSelectedTabsError do
+      render_inline(Primer::UnderlineNavComponent.new) do |c|
+        c.tab { "Tab 1" }
+        c.tab { "Tab 2" }
+        c.tab { "Tab 3" }
+      end
+    end
+
+    assert_equal("a tab must be selected", err.message)
+  end
+
+  def test_raises_if_multiple_tabs_are_selected
+    err = assert_raises Primer::TabbedComponentHelper::MultipleSelectedTabsError do
+      render_inline(Primer::UnderlineNavComponent.new) do |c|
+        c.tab(selected: true) do
+          "Tab 1"
+        end
+        c.tab { "Tab 2" }
+        c.tab(selected: true) do
+          "Tab 3"
+        end
+      end
+    end
+
+    assert_equal("only one tab can be selected", err.message)
   end
 
   def test_align_falls_back_to_default
